Fix large validation ranges by clamping range to highest row in actual data.
Fixes #1842

Some validation rules in Excel result in the sheet taking a very long time to load, because the validation loops over all numbers in the range without checking if anything might be in the targeted fields.

This fix looks at the model to find the highest actual row in use, and clamps the range to that row, before running the Range on it.